+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/SendRayRunner.php
@@ -67,7 +67,7 @@ class SendRayRunner implements StepRunnerInterface
                 $node = $this->stepProcessor->getNodeFromStep($step);
                 $nodeSettings = $this->stepProcessor->getNodeSettings($node);
 
-                $expression = '';
+                $expression = '{{input}}';
                 if (isset($nodeSettings['data']['expression'])) {
                     $expression = $nodeSettings['data']['expression'];
                 } else {
@@ -94,9 +94,7 @@ class SendRayRunner implements StepRunnerInterface
                 // phpcs:ignore PublishPressStandards.Debug.DisallowDebugFunctions.FoundRayFunction
                 $rayMessage = ray($output);
 
-                if (isset($nodeSettings['label'])) {
-                    $rayMessage->label($nodeSettings['label']);
-                }
+                $rayMessage->label($nodeSettings['label'] ?? '');
 
                 if (isset($nodeSettings['color'])) {
                     switch ($nodeSettings['color']) {