+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnSchedule.php
@@ -66,7 +66,7 @@ class OnSchedule implements StepTypeInterface
                 "fields" => [
                     [
                         "name" => "schedule",
-                        "type" => "dateOffset",
+                        "type" => "schedule",
                         "label" => __("Date offset", "post-expirator"),
                         "settings" => [
                             "hideDateSources" => [