+++ src/Modules/Workflows/Rest/RestApiV1.php
@@ -43,7 +43,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::READABLE,
                 'callback' => [$this, 'getWorkflow'],
-                'permission_callback' => [$this, 'getWorkflowPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'id' => [
                         'description' => __('The ID of the workflow', 'post-expirator'),
@@ -64,7 +64,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::READABLE,
                 'callback' => [$this, 'getWorkflowsWithManualTrigger'],
-                'permission_callback' => [$this, 'getWorkflowsWithManualTriggerPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'postType' => [
                         'description' => __('The post type', 'post-expirator'),
@@ -84,7 +84,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::CREATABLE,
                 'callback' => [$this, 'createWorkflow'],
-                'permission_callback' => [$this, 'createWorkflowPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'show_in_index' => false,
                 'show_in_rest' => true,
             ]
@@ -97,7 +97,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::EDITABLE,
                 'callback' => [$this, 'updateWorkflow'],
-                'permission_callback' => [$this, 'updateWorkflowPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'id' => [
                         'description' => __('The ID of the workflow', 'post-expirator'),
@@ -117,7 +117,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::DELETABLE,
                 'callback' => [$this, 'deleteWorkflow'],
-                'permission_callback' => [$this, 'deleteWorkflowPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'id' => [
                         'description' => __('The ID of the workflow', 'post-expirator'),
@@ -137,7 +137,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::READABLE,
                 'callback' => [$this, 'getTaxonomyTerms'],
-                'permission_callback' => [$this, 'getTaxonomyTermsPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'taxonomy' => [
                         'description' => __('The taxonomy name', 'post-expirator'),
@@ -157,7 +157,7 @@ class RestApiV1 implements RestApiManagerInterface
             [
                 'methods' => WP_REST_Server::READABLE,
                 'callback' => [$this, 'getPostWorkflowSettings'],
-                'permission_callback' => [$this, 'getPostWorkflowSettingsPermissions'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
                 'args' => [
                     'post' => [
                         'description' => __('The post ID', 'post-expirator'),
@@ -169,7 +169,21 @@ class RestApiV1 implements RestApiManagerInterface
                 'show_in_rest' => true,
             ]
         );
+
+        // Get all authors
+        register_rest_route(
+            self::BASE_PATH,
+            '/authors',
+            [
+                'methods' => WP_REST_Server::READABLE,
+                'callback' => [$this, 'getAuthors'],
+                'permission_callback' => [$this, 'checkUserCanCallApi'],
+                'show_in_index' => false,
+                'show_in_rest' => true,
+            ]
+        );
     }
+
 
     private function getWorkflowForResponse(WorkflowModel $workflowModel)
     {
@@ -349,38 +363,28 @@ class RestApiV1 implements RestApiManagerInterface
         ]);
     }
 
-    public function getWorkflowPermissions($request)
+    public function checkUserCanCallApi($request)
     {
         return current_user_can(self::PERMISSION_READ);
     }
 
-    public function createWorkflowPermissions($request)
+    public function getAuthors($request)
     {
-        return current_user_can(self::PERMISSION_CREATE);
-    }
+        $users = get_users([
+            'capability' => ['edit_posts'],
+            'orderby' => 'display_name',
+            'order' => 'ASC'
+        ]);
 
-    public function updateWorkflowPermissions($request)
-    {
-        return current_user_can(self::PERMISSION_UPDATE);
-    }
+        $users = array_map(function ($user) {
+            return [
+                'id' => $user->ID,
+                'userLogin' => $user->user_login,
+                'name' => $user->display_name,
+                'email' => $user->user_email,
+            ];
+        }, $users);
 
-    public function deleteWorkflowPermissions($request)
-    {
-        return current_user_can(self::PERMISSION_DELETE);
-    }
-
-    public function getTaxonomyTermsPermissions($request)
-    {
-        return current_user_can(self::PERMISSION_READ);
-    }
-
-    public function getPostWorkflowSettingsPermissions($request)
-    {
-        return current_user_can(self::PERMISSION_READ);
-    }
-
-    public function getWorkflowsWithManualTriggerPermissions($request)
-    {
-        return current_user_can(self::PERMISSION_READ);
+        return rest_ensure_response($users);
     }
 }